Fix resizing

+++ AsciiLib/Images/AsciiImageConverter.cs
@@ -7,8 +7,8 @@ using System.Threading.Tasks;
 
 namespace AsciiLib.Images {
     public class AsciiImageConverter {
-        private static double ratio(Bitmap b){
-            return b.Width / b.Height;
+        private static double ratio(Bitmap b) {
+            return (double)b.Width / b.Height;
         }
         public static Bitmap ResizeOutput(Bitmap original, Bitmap input) {
             double r = ratio(original);
